Move `MailTo` and `ListHeaders` classes
