+++ app/core/src/main/java/com/fsck/k9/helper/MailTo.java
