+++ app/core/src/main/java/com/fsck/k9/helper/ReplyToParser.java
@@ -12,7 +12,6 @@ import com.fsck.k9.Account;
 import com.fsck.k9.mail.Address;
 import com.fsck.k9.mail.Message;
 import com.fsck.k9.mail.Message.RecipientType;
-import com.fsck.k9.mail.internet.ListHeaders;
 
 
 public class ReplyToParser {
